limiting 60 request per seconds by client IP

+++ app/Http/Controllers/UserController.php
@@ -106,33 +106,44 @@ class UserController extends Controller
 
 				$user = $this->userService->signup($validated);
 
-				$code = random_int(1000, 9999);
+				if($user instanceof User){
 
-				$table_name = "USER_VALIDATION_TOKEN";
-				$columns = "USER_ID, VAL_TOKEN, CREATED_AT";
-				$values = $user->id.','.$code.',sysdate';
-				$result = $this->userService->insertarFila($table_name, $columns, $values);
+					$code = random_int(1000, 9999);
 
-				if ($result){
+					$table_name = "USER_VALIDATION_TOKEN";
+					$columns = "USER_ID, VAL_TOKEN, CREATED_AT";
+					$values = $user->id.','.$code.',sysdate';
+					$result = $this->userService->insertarFila($table_name, $columns, $values);
 
-					Mail::to($user->email)
-					->queue((new EmailConfirmation($user, $code))->from('ciudadanodigital@entrerios.gov.ar',
-						'Ciudadano Digital - Provincia de Entre Ríos')->subject('Validación de correo e-mail'));
-						
-					return response()->json([
-						'status' => true,
-						'message' => 'Email sent',
-					], 201);
+					if ($result){
+
+						Mail::to($user->email)
+						->queue((new EmailConfirmation($user, $code))->from('ciudadanodigital@entrerios.gov.ar',
+							'Ciudadano Digital - Provincia de Entre Ríos')->subject('Validación de correo e-mail'));
+							
+						return response()->json([
+							'status' => true,
+							'message' => 'Email sent',
+						], 201);
+
+					}else{
+
+						$user->delete();
+
+						return response()->json([
+							'status' => false,
+							'message' => 'Internal server problem, please try again later'
+						], 503);
+					}
 
 				}else{
 
-					$user->delete();
+					//users in this case is a json with error type
+					return $user;
 
-					return response()->json([
-						'status' => false,
-						'message' => 'Internal server problem, please try again later'
-					], 503);
 				}
+
+				
 
 				
 			}
@@ -158,7 +169,7 @@ class UserController extends Controller
 			$cuil=explode('/' ,$aux)[0];
 			$token= explode('/' ,$aux)[1];
 			$user = User::where('cuil', $cuil)->first();
-			
+
 			if($user){
 
 				$column_name = "USER_ID";
@@ -481,9 +492,8 @@ class UserController extends Controller
 
 		}
 	}
-    
-    
-    public function resendEmailVerificacion(CheckUserCuilRequest $request): JsonResponse
+
+	public function resendEmailVerificacion(CheckUserCuilRequest $request): JsonResponse
 	{
 		$validated = $request->validated();
 		$cuil = $validated['cuil'];
@@ -509,6 +519,7 @@ class UserController extends Controller
 					return response()->json([
 						'status' => true,
 						'message' => 'Email sent',
+						'email' => $user->email,
 					], 201);
 
 				}else{
